+++ Tests/Reqnroll.RuntimeTests/BindingSkeletons/ResourceSkeletonTemplateProviderTests.cs
@@ -20,8 +20,10 @@ namespace Reqnroll.RuntimeTests.BindingSkeletons
             var sut = new ResourceSkeletonTemplateProvider();
 
             ShouldNotBeMissing(sut.GetStepDefinitionClassTemplate(ProgrammingLanguage.CSharp));
-            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.CSharp, true));
-            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.CSharp, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.CSharp, true, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.CSharp, false, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.CSharp, true, true));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.CSharp, false, true));
         }
 
         [Fact]
@@ -30,8 +32,10 @@ namespace Reqnroll.RuntimeTests.BindingSkeletons
             var sut = new ResourceSkeletonTemplateProvider();
 
             ShouldNotBeMissing(sut.GetStepDefinitionClassTemplate(ProgrammingLanguage.VB));
-            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.VB, true));
-            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.VB, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.VB, true, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.VB, false, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.VB, true, true));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.VB, false, true));
         }
 
         [Fact]
@@ -40,8 +44,10 @@ namespace Reqnroll.RuntimeTests.BindingSkeletons
             var sut = new ResourceSkeletonTemplateProvider();
 
             ShouldNotBeMissing(sut.GetStepDefinitionClassTemplate(ProgrammingLanguage.FSharp));
-            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.FSharp, true));
-            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.FSharp, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.FSharp, true, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.FSharp, false, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.FSharp, true, true));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.FSharp, false, true));
         }
     }
 
@@ -61,8 +67,10 @@ namespace Reqnroll.RuntimeTests.BindingSkeletons
             var sut = new DefaultSkeletonTemplateProvider(new ResourceSkeletonTemplateProvider());
 
             ShouldNotBeMissing(sut.GetStepDefinitionClassTemplate(ProgrammingLanguage.CSharp));
-            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.CSharp, true));
-            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.CSharp, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.CSharp, true, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.CSharp, false, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.CSharp, true, true));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.CSharp, false, true));
         }
 
         [Fact]
@@ -71,8 +79,10 @@ namespace Reqnroll.RuntimeTests.BindingSkeletons
             var sut = new DefaultSkeletonTemplateProvider(new ResourceSkeletonTemplateProvider());
 
             ShouldNotBeMissing(sut.GetStepDefinitionClassTemplate(ProgrammingLanguage.VB));
-            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.VB, true));
-            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.VB, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.VB, true, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.VB, false, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.VB, true, true));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.VB, false, true));
         }
 
         [Fact]
@@ -81,8 +91,11 @@ namespace Reqnroll.RuntimeTests.BindingSkeletons
             var sut = new DefaultSkeletonTemplateProvider(new ResourceSkeletonTemplateProvider());
 
             ShouldNotBeMissing(sut.GetStepDefinitionClassTemplate(ProgrammingLanguage.FSharp));
-            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.FSharp, true));
-            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.FSharp, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.FSharp, true, false));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.FSharp, false, false)); 
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.FSharp, true, true));
+            ShouldNotBeMissing(sut.GetStepDefinitionTemplate(ProgrammingLanguage.FSharp, false, true));
+
         }
     }
 }